Create: - BookType
Edit:
- AuthorController > method: deleteAuthor > add tryCatch
- Author > change Column name from  id_author to author_id
- Book > add adnotations, create fields
- AuthorService > add if
- import.sql > add insert for book table

+++ src/main/java/pl/dm/libraryprojectdm/controller/AuthorController.java
@@ -39,7 +39,14 @@ public class AuthorController {
 
     @PostMapping("/delete")
     public String deleteAuthor(@RequestParam Long authorId) {
-        authorService.delete(authorId);
+        if (authorId != null) {
+            try {
+                authorService.delete(authorId);
+                return "redirect:/author/list";
+            } catch (Exception e) {
+                System.out.println("Error");
+            }
+        }
         return "redirect:/author/list";
     }
